Cleanup, added zoom and fps label, cell info box (basic)

+++ include/Types.hpp
@@ -28,8 +28,8 @@ struct CellData
         State             state;
         Side              side;
         NeighbourhoodType neighbourhoodType;
-        float             resilience = 0.0f;
-        float             fatigue    = 0.0f;
+        double            resilience = 0.0;
+        double            fatigue    = 0.0;
         bool              active     = true;
         uint8_t           stateId    = 0;
 };
